+++ src/js/components/counters.js
@@ -1,0 +1,10 @@
+import numberRollup from "@handmadeweb/number-rollup";
+
+let numberRolleups = document.querySelectorAll('.bmcb-counter');
+
+export default function () {
+  if (numberRolleups.length) {
+    console.log('yes')
+    numberRollup();
+  }
+}